refactor(theme): split monolithic theme.js into rollup-bundled modules
