+++ src/state.js
@@ -1,0 +1,44 @@
+export const app = {
+    asciiAnimationInitialized: false,
+    asciiCharData: [],
+    asciiEnabled: true,
+    tuiMode: "command",
+    results: [],
+    selected: 0,
+    lyricsObserver: null,
+    commandHistory: [],
+    commandHistoryIndex: -1,
+    playlistPanelOpen: false,
+    playlists: [],
+    playlistSongs: [],
+    playlistSongsFetchToken: 0,
+    playlistSongsFetchTimer: null,
+    selectedPlaylist: 0,
+    selectedSong: 0,
+    activePane: "playlist",
+    helpPanelOpen: false,
+    aboutPanelOpen: false,
+    themePanelOpen: false,
+    onboardingPanelOpen: false,
+    onboardingStage: "commands",
+    onboardingShowAllThemes: false,
+    lyricsPanelOpen: false,
+    lyricsLoadToken: 0,
+    lyricsActiveIndex: -1,
+    lyricsActiveLoaderIndex: -1,
+    lyricsCache: { uri: "", lines: [], synced: false, provider: "", instrumental: false, error: "" },
+    lyricsBound: false,
+    lyricsSyncInterval: null,
+    cachedLyricsRows: [],
+    cachedLyricsLoaders: [],
+    jamRole: null,
+    jamPin: null,
+    jamToken: null,
+    jamIntervalId: null,
+    jamBarPrevHidden: null,
+    jamLastAppliedUri: null,
+    themesFeedPromise: null,
+    playlistListScrollRaf: null,
+    songListScrollRaf: null,
+    navRafPending: false
+};
